structure spawn with local radius

+++ Assets/_Scripts/Core/World Generation/Structures/StructurePlacementSettings.cs
@@ -5,10 +5,10 @@ namespace HerosJourney.Core.WorldGeneration.Structures
     [Serializable]
     public struct StructurePlacementSettings
     {
-        public float radius;
+        public int radius;
         public float threshold;
 
-        public StructurePlacementSettings(float radius, float threshold)
+        public StructurePlacementSettings(int radius, float threshold)
         {
             this.radius = radius;
             this.threshold = threshold;
